fixing change event
Fixed the code for the change event in the plugin, had to manually track
key changes so the change event would work for the textboxes.  I had it
intially firing the event on every arrow movement but didn't feel that
was efficient so changed it to track and only fire change event when
text has been altered.